bajaj_new/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
python-multipart
opencv-python
